Add get property as batch function & improve typing by using dataclasses
