+++ aiowebdav2/client.py
@@ -4,21 +4,19 @@ import asyncio
 from collections.abc import AsyncIterable, Awaitable, Callable
 from datetime import datetime
 import functools
-from io import BytesIO
 import logging
 import os
 from pathlib import Path
 from re import sub
 import shutil
 from typing import IO, Any, ClassVar, Self
-from urllib.parse import unquote, urlparse, urlsplit
+from urllib.parse import unquote
 
 import aiofiles
 import aiohttp
-from lxml import etree
 
-from aiowebdav2.connection import WebDAVSettings
-from aiowebdav2.exceptions import (
+from .connection import WebDAVSettings
+from .exceptions import (
     ConnectionExceptionError,
     LocalResourceNotFoundError,
     MethodNotSupportedError,
@@ -30,8 +28,10 @@ from aiowebdav2.exceptions import (
     ResourceLockedError,
     ResponseErrorCodeError,
 )
-from aiowebdav2.typing_helper import AsyncWriteBuffer
-from aiowebdav2.urn import Urn
+from .models import Property, PropertyRequest
+from .parser import WebDavXmlUtils
+from .typing_helper import AsyncWriteBuffer
+from .urn import Urn
 
 log = logging.getLogger(__name__)
 
@@ -916,51 +916,65 @@ class Client:
         )
 
     @wrap_connection_error
-    async def get_property(self, remote_path: str, option: dict) -> Any:
+    async def get_property(
+        self, remote_path: str, requested_property: PropertyRequest
+    ) -> Property | None:
         """Get metadata property of remote resource on WebDAV server.
 
         More information you can find by link https://www.rfc-editor.org/rfc/rfc4918.html#section-9.1.
 
         :param remote_path: the path to remote resource.
-        :param option: the property attribute as dictionary with following keys:
+        :param requested_property: the property attribute as dictionary with following keys:
                        `namespace`: (optional) the namespace for XML property which will be set,
                        `name`: the name of property which will be set.
         :return: the value of property or None if property is not found.
         """
+        result = await self.get_properties(remote_path, [requested_property])
+        return result[0] if result and result[0] else None
+
+    @wrap_connection_error
+    async def get_properties(
+        self, remote_path: str, requested_properties: list[PropertyRequest]
+    ) -> list[Property]:
+        """Get metadata properties of remote resource on WebDAV server."""
         urn = Urn(remote_path)
         if not await self.check(urn.path()):
             raise RemoteResourceNotFoundError(urn.path())
 
-        data = WebDavXmlUtils.create_get_property_request_content(option)
+        data = WebDavXmlUtils.create_get_property_batch_request_content(
+            requested_properties
+        )
         response = await self.execute_request(
             action="get_property", path=urn.quote(), data=data
         )
-        return WebDavXmlUtils.parse_get_property_response(
-            await response.read(), option["name"]
+        return WebDavXmlUtils.parse_get_properties_response(
+            await response.read(), requested_properties
         )
 
     @wrap_connection_error
-    async def set_property(self, remote_path: str, option: list) -> None:
+    async def set_property(self, remote_path: str, prop: Property) -> None:
         """Set metadata property of remote resource on WebDAV server.
 
         More information you can find by link https://www.rfc-editor.org/rfc/rfc4918.html#section-9.2.
 
         :param remote_path: the path to remote resource.
-        :param option: the property attribute as dictionary with following keys:
+        :param prop: the property attribute as dictionary with following keys:
                        `namespace`: (optional) the namespace for XML property which will be set,
                        `name`: the name of property which will be set,
                        `value`: (optional) the value of property which will be set. Defaults is empty string.
         """
-        await self.set_property_batch(remote_path=remote_path, option=[option])
+        await self.set_property_batch(remote_path=remote_path, properties=[prop])
 
     @wrap_connection_error
-    async def set_property_batch(self, remote_path: str, option: list) -> None:
+    async def set_property_batch(
+        self, remote_path: str, properties: list[Property]
+    ) -> None:
         """Set batch metadata properties of remote resource on WebDAV server in batch.
 
         More information you can find by link https://www.rfc-editor.org/rfc/rfc4918.html#section-9.2.
 
         :param remote_path: the path to remote resource.
-        :param option: the property attributes as list of dictionaries with following keys:
+        :param properties: the property attributes as list of dictionaries with following keys:
                        `namespace`: (optional) the namespace for XML property which will be set,
                        `name`: the name of property which will be set,
                        `value`: (optional) the value of property which will be set. Defaults is empty string.
@@ -969,7 +983,7 @@ class Client:
         if not await self.check(urn.path()):
             raise RemoteResourceNotFoundError(urn.path())
 
-        data = WebDavXmlUtils.create_set_property_batch_request_content(option)
+        data = WebDavXmlUtils.create_set_property_batch_request_content(properties)
         await self.execute_request(action="set_property", path=urn.quote(), data=data)
 
     @wrap_connection_error
@@ -1269,261 +1283,19 @@ class Resource:
         """Unpublish the resource."""
         return await self.client.unpublish(self.urn.path())
 
-    async def get_property(self, option: dict[str, Any]) -> str | None:
+    async def get_property(
+        self, requested_property: PropertyRequest
+    ) -> Property | None:
         """Get metadata property of the resource."""
         return await self.client.get_property(
-            remote_path=self.urn.path(), option=option
+            remote_path=self.urn.path(), option=requested_property
         )
 
-    async def set_property(self, option: dict[str, Any], value: Any) -> None:
+    async def set_property(self, name: str, value: str, namespace: str = "") -> None:
         """Set metadata property of the resource."""
-        option["value"] = value.__str__()
-        await self.client.set_property(remote_path=self.urn.path(), option=option)
-
-
-class WebDavXmlUtils:
-    """WebDAV XML utils."""
-
-    @staticmethod
-    def parse_get_list_info_response(content: bytes) -> list[dict]:
-        """Parse of response content XML from WebDAV server and extract file and directory infos.
-
-        :param content: the XML content of HTTP response from WebDAV server for getting list of files by remote path.
-        :return: list of information, the information is a dictionary and it values with following keys:
-                 `created`: date of resource creation,
-                 `name`: name of resource,
-                 `size`: size of resource,
-                 `modified`: date of resource modification,
-                 `etag`: etag of resource,
-                 `content_type`: content type of resource,
-                 `isdir`: type of resource,
-                 `path`: path of resource.
-        """
-        try:
-            tree = etree.fromstring(content)
-            infos = []
-            for response in tree.findall(".//{DAV:}response"):
-                href_el = next(iter(response.findall(".//{DAV:}href")), None)
-                if href_el is None:
-                    continue
-                path = unquote(urlsplit(href_el.text).path)
-                is_dir = len(response.findall(".//{DAV:}collection")) > 0
-                info = WebDavXmlUtils.get_info_from_response(response)
-                info["isdir"] = is_dir
-                info["path"] = path
-                infos.append(info)
-        except etree.XMLSyntaxError:
-            return []
-        else:
-            return infos
-
-    @staticmethod
-    def parse_get_list_response(content: bytes) -> list[Urn]:
-        """Parse of response content XML from WebDAV server and extract file and directory names.
-
-        :param content: the XML content of HTTP response from WebDAV server for getting list of files by remote path.
-        :return: list of extracted file or directory names.
-        """
-        try:
-            tree = etree.fromstring(content)
-            urns = []
-            for response in tree.findall(".//{DAV:}response"):
-                href_el = next(iter(response.findall(".//{DAV:}href")), None)
-                if href_el is None:
-                    continue
-                href = Urn.separate + unquote(urlsplit(href_el.text).path)
-                is_dir = len(response.findall(".//{DAV:}collection")) > 0
-                urns.append(Urn(href, directory=is_dir))
-        except etree.XMLSyntaxError:
-            return []
-        else:
-            return urns
-
-    @staticmethod
-    def create_free_space_request_content() -> bytes:
-        """Create an XML for requesting of free space on remote WebDAV server.
-
-        :return: the XML string of request content.
-        """
-        root = etree.Element("propfind", xmlns="DAV:")
-        prop = etree.SubElement(root, "prop")
-        etree.SubElement(prop, "quota-available-bytes")
-        etree.SubElement(prop, "quota-used-bytes")
-        tree = etree.ElementTree(root)
-        return WebDavXmlUtils.etree_to_string(tree)
-
-    @staticmethod
-    def parse_free_space_response(content: bytes, hostname: str) -> int | None:
-        """Parse of response content XML from WebDAV server and extract an amount of free space.
-
-        :param content: the XML content of HTTP response from WebDAV server for getting free space.
-        :param hostname: the server hostname.
-        :return: an amount of free space in bytes.
-        """
-        try:
-            tree = etree.fromstring(content)
-            node = tree.find(".//{DAV:}quota-available-bytes")
-            if node is not None:
-                return int(node.text)
-            raise MethodNotSupportedError(name="free", server=hostname)
-        except TypeError as err:
-            raise MethodNotSupportedError(name="free", server=hostname) from err
-        except etree.XMLSyntaxError:
-            return None
-
-    @staticmethod
-    def get_info_from_response(response: etree.Element) -> dict:
-        """Get information attributes from response.
-
-        :param response: XML object of response for the remote resource defined by path
-        :return: a dictionary of information attributes and them values with following keys:
-                 `created`: date of resource creation,
-                 `name`: name of resource,
-                 `size`: size of resource,
-                 `modified`: date of resource modification,
-                 `etag`: etag of resource,
-                 `content_type`: content type of resource.
-        """
-        find_attributes = {
-            "created": ".//{DAV:}creationdate",
-            "name": ".//{DAV:}displayname",
-            "size": ".//{DAV:}getcontentlength",
-            "modified": ".//{DAV:}getlastmodified",
-            "etag": ".//{DAV:}getetag",
-            "content_type": ".//{DAV:}getcontenttype",
-        }
-        info = {}
-        for name, value in find_attributes.items():
-            info[name] = response.findtext(value)
-        return info
-
-    @staticmethod
-    def parse_info_response(content: bytes, path: str, hostname: str) -> dict:
-        """Parse of response content XML from WebDAV server and extract an information about resource.
-
-        :param content: the XML content of HTTP response from WebDAV server.
-        :param path: the path to resource.
-        :param hostname: the server hostname.
-        :return: a dictionary of information attributes and them values with following keys:
-                 `created`: date of resource creation,
-                 `name`: name of resource,
-                 `size`: size of resource,
-                 `modified`: date of resource modification,
-                 `etag`: etag of resource,
-                 `content_type`: content type of resource.
-        """
-        response = WebDavXmlUtils.extract_response_for_path(
-            content=content, path=path, hostname=hostname
+        await self.client.set_property(
+            self.urn.path(), Property(name=name, value=value, namespace=namespace)
         )
-        return WebDavXmlUtils.get_info_from_response(response)
-
-    @staticmethod
-    def parse_is_dir_response(content: bytes, path: str, hostname: str) -> bool:
-        """Parse of response content XML from WebDAV server and extract an information about resource.
-
-        :param content: the XML content of HTTP response from WebDAV server.
-        :param path: the path to resource.
-        :param hostname: the server hostname.
-        :return: True in case the remote resource is directory and False otherwise.
-        """
-        response = WebDavXmlUtils.extract_response_for_path(
-            content=content, path=path, hostname=hostname
-        )
-        resource_type = response.find(".//{DAV:}resourcetype")
-        if resource_type is None:
-            raise MethodNotSupportedError(name="is_dir", server=hostname)
-        dir_type = resource_type.find("{DAV:}collection")
-
-        return dir_type is not None
-
-    @staticmethod
-    def create_get_property_request_content(option: dict) -> bytes:
-        """Create an XML for requesting of getting a property value of remote WebDAV resource.
-
-        :param option: the property attributes as dictionary with following keys:
-                       `namespace`: (optional) the namespace for XML property which will be get,
-                       `name`: the name of property which will be get.
-        :return: the XML string of request content.
-        """
-        root = etree.Element("propfind", xmlns="DAV:")
-        prop = etree.SubElement(root, "prop")
-        etree.SubElement(
-            prop, option.get("name", ""), xmlns=option.get("namespace", "")
-        )
-        tree = etree.ElementTree(root)
-        return WebDavXmlUtils.etree_to_string(tree)
-
-    @staticmethod
-    def parse_get_property_response(content: bytes, name: str) -> str:
-        """Parse of response content XML from WebDAV server for getting metadata property value for some resource.
-
-        :param content: the XML content of response as string.
-        :param name: the name of property for finding a value in response
-        :return: the value of property if it has been found or None otherwise.
-        """
-        tree = etree.fromstring(content)
-        return tree.xpath("//*[local-name() = $name]", name=name)[0].text
-
-    @staticmethod
-    def create_set_property_batch_request_content(options: list[dict]) -> bytes:
-        """Create an XML for requesting of setting a property values for remote WebDAV resource in batch.
-
-        :param options: the property attributes as list of dictionaries with following keys:
-               `namespace`: (optional) the namespace for XML property which will be set,
-               `name`: the name of property which will be set,
-               `value`: (optional) the value of property which will be set. Defaults is empty string.
-        :return: the XML string of request content.
-        """
-        root_node = etree.Element("propertyupdate", xmlns="DAV:")
-        set_node = etree.SubElement(root_node, "set")
-        prop_node = etree.SubElement(set_node, "prop")
-        for option in options:
-            opt_node = etree.SubElement(
-                prop_node, option["name"], xmlns=option.get("namespace", "")
-            )
-            opt_node.text = option.get("value", "")
-        tree = etree.ElementTree(root_node)
-        return WebDavXmlUtils.etree_to_string(tree)
-
-    @staticmethod
-    def etree_to_string(tree: etree.ElementTree) -> bytes:
-        """Create string from lxml.etree.ElementTree with XML declaration and UTF-8 encoding.
-
-        :param tree: the instance of ElementTree
-        :return: the string of XML.
-        """
-        buff = BytesIO()
-        tree.write(buff, xml_declaration=True, encoding="UTF-8")
-        return buff.getvalue()
-
-    @staticmethod
-    def extract_response_for_path(
-        content: bytes, path: str, hostname: str
-    ) -> etree.Element:
-        """Extract single response for specified remote resource.
-
-        :param content: raw content of response as string.
-        :param path: the path to needed remote resource.
-        :param hostname: the server hostname.
-        :return: XML object of response for the remote resource defined by path.
-        """
-        prefix = urlparse(hostname).path
-        try:
-            tree = etree.fromstring(content)
-            responses = tree.findall("{DAV:}response")
-            n_path = Urn.normalize_path(path)
-
-            for resp in responses:
-                href = resp.findtext("{DAV:}href")
-
-                if Urn.compare_path(n_path, href) is True:
-                    return resp
-                href_without_prefix = href.removeprefix(prefix)
-                if Urn.compare_path(n_path, href_without_prefix) is True:
-                    return resp
-            raise RemoteResourceNotFoundError(path)
-        except etree.XMLSyntaxError as err:
-            raise MethodNotSupportedError(name="is_dir", server=hostname) from err
 
 
 class LockClient(Client):
