+++ pyproject.toml
@@ -124,6 +124,7 @@ ignore-imports = true
 
 [tool.pylint.FORMAT]
 max-line-length = 120
+extension-pkg-allow-list = "lxml"
 
 [tool.pytest.ini_options]
 addopts = "--cov"
