--- tests/responses/free_space_incorrect.xml
@@ -1,9 +1,0 @@
-12321dadadasdasd<?xml version="1.0" encoding="utf-8"?>
-<d:multistatus xmlns:d="DAV:">
-    <d:response>
-        <d:href>/</d:href>
-        <d:propstat>
-            <d:status>HTTP/1.1 200 OK</d:status>
-        </d:propstat>
-    </d:response>
-</d:multistatus>
