--- tests/responses/get_list_incorrect.xml
@@ -1,50 +1,0 @@
-asdadasDASAS<?xml version="1.0" encoding="utf-8"?>
-<D:multistatus xmlns:D="DAV:">
-
-    <D:response xmlns:lp1="DAV:" xmlns:lp2="http://apache.org/dav/props/">
-        <D:href>/test_dir/test.txt</D:href>
-        <D:propstat>
-            <D:prop>
-                <lp1:resourcetype/>
-                <lp1:creationdate>2020-04-10T21:59:43Z</lp1:creationdate>
-                <lp1:getcontentlength>
-                    41
-                </lp1:getcontentlength>
-                <lp1:getlastmodified>Fri, 10 Apr 2020 21:59:43 GMT</lp1:getlastmodified>
-                <lp1:getetag>"29-5a2f6d9cf8d39"</lp1:getetag>
-                <lp2:executable>
-                    F
-                </lp2:executable>
-                <D:supportedlock>
-                    <D:lockentry>
-                        <D:lockscope>
-                            <D:exclusive/>
-                        </D:lockscope>
-
-                        <D:locktype>
-                            <D:write/>
-                        </D:locktype>
-
-                    </D:lockentry>
-
-                    <D:lockentry>
-                        <D:lockscope>
-                            <D:shared/>
-                        </D:lockscope>
-
-                        <D:locktype>
-                            <D:write/>
-                        </D:locktype>
-
-                    </D:lockentry>
-
-                </D:supportedlock>
-                <D:lockdiscovery/>
-                <D:getcontenttype>text/plain</D:getcontenttype>
-            </D:prop>
-            <D:status>HTTP/1.1 200 OK</D:status>
-        </D:propstat>
-
-    </D:response>
-
-</D:multistatus>
